requirements.txt
python-dotenv==1.0.0
pyairtable==2.1.0
supabase==1.0.3
requests==2.31.0
python-dateutil==2.8.2 